Move OCI copy tests to dedicated file
